Change string in French translation
This is an opinionated suggestion. The current French string for "Specify where to open the linked document." is a straight translation. It is grammatically correct, but feels very technical, very "W3C spec". If this was intended then you can ignore this PR :)

The string I suggest is how actual end users, clients, would phrase it 99% of the times. They would always say "open the link" ("Ouvrir le lien") instead of "Open the linked document" ("Ouvrir le document lié"). I also use the infinitive tense instead of imperative, to align with how Kirby core is translated.