feat: add support for blocks and compatibility to Kirby 3.7
Closes #59

BREAKING CHANGE: While adding support for Kirby blocks the user interface and the user experience got a revision that itself led to a major rework of the code base.

I simplified the process of selecting pages and optimised the configurability. It is now possible not only to customize the page title, but also to sort and filter the list of results.